Temp input from file added.

+++ app/src/main/java/com/ignation/speisefant/domain/Product.kt
@@ -3,8 +3,8 @@ package com.ignation.speisefant.domain
 data class Product(
     val name: String,
     val image: String,
-    val newPrice: Float,
-    val oldPrice: Float,
+    val newPrice: Int,
+    val oldPrice: Int,
     val shop: String,
     val startDate: String,
     val endDate: String,
